no rolling avg

+++ Masquerade4997Library/src/main/java/Library4997/MasqSensors/MasqVoltageSensor.java
@@ -27,5 +27,5 @@ public class MasqVoltageSensor {
         return sum/num;
     }
 
-    public double getVoltageInstantaneous() {return (int) sensor.getVoltage();}
+    public double getVoltageInstantaneous() {return sensor.getVoltage();}
 }
